refactor : moved MatrixMultCuBLAS member functions to cpp
