+++ gemm/CMakeLists.txt
@@ -14,7 +14,8 @@ find_package(CUDAToolkit REQUIRED)
 
 add_library(gemm SHARED
     ${CMAKE_SOURCE_DIR}/src/Matrix.cpp
-    ${CMAKE_SOURCE_DIR}/src/MatrixMult.cpp)
+    ${CMAKE_SOURCE_DIR}/src/MatrixMult.cpp
+    ${CMAKE_SOURCE_DIR}/src/MatrixMultCuBLAS.cpp)
 
 target_include_directories(gemm PUBLIC ${CMAKE_SOURCE_DIR}/include)
